requirements.txt
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-filebrowser==3.6.1
django-grappelli==2.7.1
django-herokuapp==0.9.20
django-require==1.0.8
django-require-s3==1.0.0
django-storages==1.1.8
django-toolbelt==0.0.1
gunicorn==19.3.0
Pillow==3.0.0
psycopg2==2.6.1
pytz==2015.4
sh==1.11
static3==0.6.1
waitress==0.8.9
wheel==0.24.0
